Bug 702431: Don't cast a unsigned long to an unsigned int
When allocating the pixmap for a display device, we were casting the unsigned
long bitmap size to an unsigned int, meaning on LP64 (where a long is 64 bits)
we could end up truncating the value.

Change it to cast to a size_t.

Ultimately, we'll want to change the display device structure to store a size_t
rather than unsigned long, but that will mean changing the display device's
public API.

+++ devices/gdevdsp.c
@@ -1457,7 +1457,7 @@ display_alloc_bitmap(gx_device_display * ddev, gx_device * param_dev)
     }
     else {
         ddev->pBitmap = gs_alloc_byte_array_immovable(ddev->memory->non_gc_memory,
-                (uint)ddev->ulBitmapSize, 1, "display_alloc_bitmap");
+                (size_t)ddev->ulBitmapSize, 1, "display_alloc_bitmap");
     }
 
     if (ddev->pBitmap == NULL) {
